requirements.txt
Pillow
pyyaml
importlib_resources; python_version < '3.8'
